+++ frontend/lib/pages/checkout.dart
@@ -1,108 +1,197 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_stripe/flutter_stripe.dart';
+import 'package:tripsitter/classes/payment.dart';
+import 'package:tripsitter/classes/profile.dart';
+import 'package:tripsitter/classes/trip.dart';
+import 'package:tripsitter/components/navbar.dart';
+import 'package:tripsitter/components/payment.dart';
+import 'package:tripsitter/helpers/api.dart';
 
 class CheckoutPage extends StatefulWidget {
-  const CheckoutPage({super.key});
+  final Trip trip;
+  final List<UserProfile> profiles;
+  const CheckoutPage({required this.trip, required this.profiles, super.key});
 
   @override
-  State<CheckoutPage> createState() => _LoginPageState();
+  State<CheckoutPage> createState() => _CheckoutPageState();
 }
 
-class _LoginPageState extends State<CheckoutPage> {
-  final FirebaseAuth _auth = FirebaseAuth.instance;
+class _CheckoutPageState extends State<CheckoutPage> {
 
-  User? _user;
+  Trip get trip => widget.trip;
+  CardFieldInputDetails? cardDetails;
 
-  Map<String, dynamic>? paymentIntent;
-
-  createPaymentIntent() async {
-    try{
-      // Connect to Stripe API from backend
-      return null;
-    } catch (e) {
-      throw Exception(e.toString());
-    }
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    _auth.authStateChanges().listen((event) {
-      setState(() {
-        _user = event;
-      });
-      if (_user == null) {
-        print('User is currently signed out!');
-      } else {
-        print('User is signed in!');
-      }
-    });
-  }
+  bool loading = false;
 
   @override
   Widget build(BuildContext context) {
+    User? user = FirebaseAuth.instance.currentUser;
+    if(user == null) {
+      return const Center(
+        child: CircularProgressIndicator()
+      );
+    }
     return Scaffold(
-      appBar: AppBar(
-        title: const Text('Checkout'),
-      ),
-      body: _user != null ? _checkoutButton() : _returnButton(),
-      );
-  }
-
-  Widget _returnButton() {
-    return Center(
-        child: Column(
-          children: [
-            const Center(
-              child: Text("You are not Signed In!"),
+      appBar: const TripSitterNavbar(),
+      body: AbsorbPointer(
+        absorbing: loading,
+        child: Center(
+          child: Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: ConstrainedBox(
+              constraints: const BoxConstraints(
+                maxWidth: 600,
+              ),
+              child: ListView(
+                children: [
+                  Text("Checkout Page"),
+                  Text("Trip: ${widget.trip.name}"),
+                  if(trip.flights.isNotEmpty)
+                    ...[
+                      Text("Flights: \$${trip.flightsPrice} total"),
+                      for(var flight in trip.flights)
+                        ...[
+                          Text("${flight.departureAirport} -> ${flight.arrivalAirport} (${flight.price == null ? "Unknown price" : "\$${flight.price} total"})"),
+                          Text(flight.members.map((e) => widget.profiles.firstWhere((profile) => profile.id == e).name).join(", ")),
+                        ],
+                        Container(height: 10)
+                    ],
+                  if(trip.hotels.isNotEmpty)
+                    ...[
+                      Text("Hotels: \$${trip.hotelsPrice} total"),
+                      for(var hotel in trip.hotels)
+                        ...[
+                          Text("${hotel.name} (${hotel.price == null ? "Unknown price" : "\$${hotel.price} total"})"),
+                          Text(hotel.members.map((e) => widget.profiles.firstWhere((profile) => profile.id == e).name).join(", ")),
+                        ],
+                        Container(height: 10)
+                    ],
+                  if(trip.rentalCars.isNotEmpty)
+                    ...[
+                      Text("Rental Cars: \$${trip.rentalCarsPrice} total"),
+                      for(var car in trip.rentalCars)
+                        ...[
+                          Text("${car.name} (\$${car.price} total)"),
+                          Text(car.members.map((e) => widget.profiles.firstWhere((profile) => profile.id == e).name).join(", ")),
+                        ],
+                        Container(height: 10)
+                    ],
+                  if(trip.activities.isNotEmpty)
+                    ...[
+                      Text("Activities: \$${trip.activitiesPrice} total"),
+                      for(var activity in trip.activities)
+                        ...[
+                          Text("${activity.event.name} (${activity.price == null ? "Unknown price" : "\$${activity.price} total"})"),
+                          Text(activity.participants.map((e) => widget.profiles.firstWhere((profile) => profile.id == e).name).join(", ")),
+                        ],
+                        Container(height: 10)
+                    ],
+                  Text("Total price: \$${trip.totalPrice}"),
+                  ConstrainedBox(
+                    constraints: const BoxConstraints(
+                      maxHeight: 100,
+                    ),
+                    child: CardField(
+                      key: Key("payment-${trip.id}"),
+                      onCardChanged: (CardFieldInputDetails? card) {
+                        if(card != null) {
+                          setState(() {
+                            cardDetails = card;
+                          });
+                        }
+                      },    
+                    ),
+                  ),
+                  if(cardDetails?.complete ?? false)
+                    ElevatedButton(
+                      onPressed: () async {
+                        setState(() {
+                          loading = true;
+                        });
+                        PaymentIntentData data = await TripsitterApi.createPaymentIntent(user.uid, trip);
+                        print("Payment intent data created ${data.clientSecret}");
+                        try {
+                          PaymentIntent intent = await Stripe.instance.confirmPayment(
+                            paymentIntentClientSecret: data.clientSecret,
+                            data: PaymentMethodParams.card(paymentMethodData: PaymentMethodData(
+                              billingDetails: BillingDetails(
+                                name: user.displayName,
+                                email: user.email,
+                              ),
+                            )),
+                          );
+                          print("Intent processed");
+                          print(intent.amount);
+                          print(intent.receiptEmail);
+                          print(intent.status);
+          
+                          if(intent.status == PaymentIntentsStatus.Succeeded) {
+                            print("PAYMENT COMPLETE!");
+                            if(mounted) {
+                              setState(() {
+                                loading = false;
+                              });
+                            
+                            }
+                          }
+                          else if(mounted){
+                            await showDialog(
+                              context: context, 
+                              builder: (context) => AlertDialog(
+                                title: const Text("Payment failed"),
+                                content: const Text("Payment failed, please try again"),
+                                actions: [
+                                  TextButton(
+                                    onPressed: () {
+                                      Navigator.of(context).pop();
+                                    }, 
+                                    child: const Text("OK")
+                                  ),
+                                ],
+                              )
+                            );
+                            if(mounted) {
+                              setState(() {
+                                loading = false;
+                              });
+                            
+                            }
+                          }
+                        } catch (e) {
+                          print("Error creating payment intent");
+                          print(e);
+                          await showDialog(
+                            context: context, 
+                            builder: (context) => AlertDialog(
+                              title: const Text("Payment failed"),
+                              content: const Text("Payment failed, please try again"),
+                              actions: [
+                                TextButton(
+                                  onPressed: () {
+                                    Navigator.of(context).pop();
+                                  }, 
+                                  child: const Text("OK")
+                                ),
+                              ],
+                            )
+                          );
+                          if(mounted) {
+                            setState(() {
+                              loading = false;
+                            });
+                          }
+                        }
+                        
+                      },
+                      child: loading ? const CircularProgressIndicator() : const Text("Purchase Trip!"),
+                    ),
+                ],
+              ),
             ),
-            const SizedBox(height: 20,),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.pushNamed(context, "/");
-              },
-              child: const Text("Return to Home"),),
-          ],
-        )
-    );
-  }
-
-  Widget _checkoutButton() {
-    return Center(child: SizedBox(
-      height: 50,
-      child: ElevatedButton(
-        onPressed: makePayment, child: const Text("Checkout"),
+          ),
+        ),
       ),
-    ),);
-  }
-
-  void displayPaymentSheet () async{
-    try {
-      await Stripe.instance.presentPaymentSheet();
-    } catch (e) {
-      print("Payment sheet display failed");
-    }
-  }
-
-  void makePayment () async {
-    try {
-      paymentIntent = await createPaymentIntent();
-
-      var googlePay = const PaymentSheetGooglePay(
-          merchantCountryCode: "US",
-        currencyCode: "US",
-        testEnv: true,
-      );
-      Stripe.instance.initPaymentSheet(paymentSheetParameters: SetupPaymentSheetParameters(
-        paymentIntentClientSecret: paymentIntent!["client_secret"],
-        style: ThemeMode.dark,
-        merchantDisplayName: "TripSitter",
-        googlePay: googlePay,
-      ));
-      displayPaymentSheet();
-    } catch (e) {
-      throw Exception(e.toString());
-    }
+    );
   }
 }
